web-application/webapp/forms.py
from django import forms
from django.utils.translation import ugettext_lazy as _
from snowpenguin.django.recaptcha2.fields import ReCaptchaField
from snowpenguin.django.recaptcha2.widgets import ReCaptchaWidget

from webapp.models import Feedback


class FeedbackForm(forms.ModelForm):
    class Meta:
        model = Feedback
        fields = ('name', 'phone', 'email', 'message')

    captcha = ReCaptchaField(widget=ReCaptchaWidget())
